tasks/side-channel/requirements.txt
absl-py==0.1.10
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
h5py==2.7.1
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.4
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.0
Pillow==5.0.0
protobuf==3.5.1
PyYAML==3.12
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
urllib3==1.22
Werkzeug==0.14.1
